fix: remove postal code module
