+++ src/controllers/address.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { validateCountry, cityExistsForCountry, postalCodeExistsForCity, stateExistsForCountry } from '../services';
+import { validateCountry, cityExistsForCountry, postalCodeExistsForDistrict, stateExistsForCountry } from '../services';
 import { sendApiResponse, sendErrorResponse } from '../utils';
 
 export async function validateAddressController(
@@ -10,7 +10,7 @@ export async function validateAddressController(
     const { country, city, postalCode, state } = req.body;
     const validatedAddress = await validateCountry(country);
     const isValidCity = cityExistsForCountry(country, city);
-    const isValidPostalCode = postalCodeExistsForCity(country, city, postalCode);
+    const isValidPostalCode = postalCodeExistsForDistrict(country, city, postalCode);
     const isValidState = stateExistsForCountry(country, state);
 
     if (!isValidCity || !isValidPostalCode || !isValidState) {
