+++ src/controllers/index.ts
@@ -1,6 +1,5 @@
 export * from './address';
 export * from './city';
-export * from './postalCode';
 export * from './state';
 export * from './country';
 export * from './district';
