+++ src/services/district.ts
@@ -9,3 +9,12 @@ export function districtExistsForCountry(countryCode: string, districtName: stri
   const districts = getDistrictsByCountry(countryCode);
   return districts.some(district => district.name === districtName);
 }
+
+export function getDistrictsByPostalCode(postalCode: string): District[] | [] {
+  return districts.filter(district => district?.postalCode === postalCode);
+}
+
+export function postalCodeExistsForDistrict(countryCode: string, districtName: string, postalCode: string): boolean {
+  const districts = getDistrictsByCountry(countryCode);
+  return districts.some(district => district.name === districtName && district.postalCode === postalCode);
+}
