+++ src/services/index.ts
@@ -1,5 +1,4 @@
 export * from './country';
 export * from './city';
-export * from './postalCode';
 export * from './state';
 export * from './district';
